requirements.txt
# Core LLM & Agent Framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0

# Vector Store & Embeddings
pinecone-client>=3.0.0
sentence-transformers>=2.7.0
tiktoken>=0.7.0

# PDF Processing
PyMuPDF>=1.24.0
pypdf>=4.0.0

# Web Search
tavily-python>=0.3.0

# Financial Calculations
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0

# UI
streamlit>=1.35.0
plotly>=5.22.0

# PDF Generation (for sample statements)
reportlab>=4.2.0
fpdf2>=2.7.0

# Evaluation
ragas>=0.1.0
datasets>=2.19.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
tenacity>=8.3.0
loguru>=0.7.0

# Development
jupyter>=1.0.0
ipykernel>=6.29.0
